Game initiation and move validation logic
The server waits for at least one player to connect and sends them a request to initiate the game. After that, the server waits for the connection of the second player (if not already connected) and sends them a request to initiate the game as well. Once confirmed by the players, the game begins. The server only accepts one move from a player at a time. There is also a check for the correctness of the player's move, whether it is a valid number and if it has been already used. If the move is not a number or the number has already been used, the server sends a request for a new move until the player meets the required conditions. At any point during the game, each player can type "stop" to interrupt the game, and they will be considered as the loser while their opponent receives a message about winning.